Hot sort for the Files page
Now when you pick a new sort order from the "Sort files by:" menu on the Files page, all expanded directories are refreshed in the new sort order. Previously you had to collapse and re-expand a directory to see the new sort order.
Affected:
 - intramine_filetree.pl
 - files.js

Also, four small changes to JS files to avoid an 'undefined' nuisance.
And the JS corrections are in:
 - cmViewerStart.js
 - editor.js
 - status.js
 - tooltip.js