Minor Search bug fixed, ToDo now flashes
If all but a few languages or extensions were selected on the Search form (eg everything but Plain Text) the exclusions were being ignored. This is now fixed.

And the "ToDo" item in the navigation bar on any open browser page now flashes if ToDo items are changed anywhere. Fairly trivial, but the main goal is to show how to implement an SSE in IntraMine that reaches out to all web pages.
Affected:
 - all 13 servers with a Nav bar now include todoFlash.js
 - intramine_todolist.pl now broadcasts a 'todoFlash' event when saving ToDo data
 - todoFlash.js listens for 'todoFlash' events, and does the flashing.
Search the IntraMine folder for 'todoFlash' to see the related code.