IMPORTANT update with bug fix
Recently I accidentally disabled almost all intranet autolinking. Somewhat ironic, since the main unique feature of IntraMine is intended to be intranet autolinking. Anyway, autolinking is back. This update is worth getting regardless, since it also has many small improvements.

The main headline here was supposed to be the completely redone POD viewer, but now I have to mention incidentally that it's included. The version in this update starts with Pod::Simple::HTML and converts its output to Gloss (IntraMine's markdown variant), which is then displayed by IntraMine's viewer for text files (which does Gloss). A bit roundabout, but now POD display has line numbers, a Toggle button, document-wide selection highlighting in the text and down the scroll bar, a table of contents down the left instead of at the top so it never disappears, and other lesser but nice things such as a "true size" yellow thumb in the scroll bar for large documents. And autolinking:)

Affected: many files. gloss2html.pm is new.