The ToDo service no longer uses jQuery
The JavaScript for the ToDo service is now vanilla JavaScript (using Dragula for drag-and-drop, and Vanilla JS Datepicker for a date picking calendar), and one hopes it will be easier to read and maintain. All features including the ToDo.txt file format have been retained, and there's now a "ghost" image of a dragged task that shows where it will end up when you drop it. If you're updating, restart the ToDo service (or IntraMine) after grabbing the update, and refresh any open IntraMine windows.
Affected:
 - todo.js has been completely rewritten
 - todo.js has been completely rewritten
 - todoGetPutData.js several changes
 - intramine_todolist.pl minor changes
 - datepicker and dragula JS and CSS files added.