Link resolution in Perl file commments changed
In Perl file comments, external links now take precedence over internal links in Perl files, where a sub is defined in the current file but an external link is provided. Eg if "sub quickSort()" is defined in a file then quickSort() in a comment will link to that internal definition, but someOtherFile.pl#quickSort() will link to the definition in someOtherFile.pl.
Affected: intramine_viewer.pl#AddInternalLinksToPerlLine().