Automatic directory linking!
Type a directory name or part of a path to one within quotes in a text file or source file comment, eg "Change Logs" or "proj99/change logs", and it will be turned into a link when displayed using IntraMine's Viewer or Editor. For example, "Change Logs" will be a link to the /Change Logs/ directory that's closest to the file you typed it in. Click on a directory link and a new Files tab opens in your browser showing an expanded view of the directory. Restrictions on the directory path you type are at the theoretical minimum, as with automatic file links (only additional directories that remove ambiguity are needed, the matching directory closest to where you're typing is preferred, directories don't have to be in order, use forward or back slashes between dir names, case insensitive, etc).There's more in the README below near the end of the Autolinks section.