Editor restores unsaved changes
Previously, IntraMine's Editor would warn you in a dialog to Cancel when closing without saving changes, if you in fact wanted to keep the changes. In this update, unsaved edits are saved behind the scenes, with no warning if you close without saving, but the edits are restored the next time you open the file, with a note up top that you should Undo if you want to discard the changes permanently. This new approach works even if your PC shuts down unexpectedly, whereas the previous approach didn't.