Blank line required before ATX headings
These are headings that start with '#'. There must now be a blank line before the # heading line for it to count as a heading. This affects IntraMine's Viewer, gloss2html.pl, and the ToDo page. If you don't want that: download Intramine, restart it, open data/intramine_config_6.txt, and change the value of HASH_HEADING_NEEDS_BLANK_BEFORE to 0. Then restart again.