Minor link improvements
The Linker was rejecting quoted file names too often if surrounding text suggested an HTML context. Potential links are now not rejected if their text contains a period, suggesting a file extension.
And thrown in with this little update, FLASH links now accept ':' in addition to '#' as the start of an anchor name or line number, as in tooltip.js:321 or intramine_linker.pl:ShortenedClassAnchor(). This is handy with error summaries copied from Chrome's console.