Faster Files page, and some video links
In the Viewer and Editor and also the Files page, video links are supported, though only on the PC running IntraMine. gloss2html.pl also supports video links in the generated HTML files, but the movies must accompany the HTML (and thus with videos the HTML documents are not "standalone"). For details see 'Gloss.html" and "gloss2html.pl for standalone Gloss files.txt" in the Documentation folder.

And the Files page is now much faster, pretty much as fast as Windows File Explorer.