Editor spell check for .txt
IntraMine's Editor can now check spelling in .txt files: you'll see a new "Check" button for that at top right. This is an on/off toggle, and reads "Checking" when spell checking is enabled. Even though it's checking against a list of 467,000 words you won't notice any slowdown. On/off state is persistent on a per-file basis for each user. "Errors" are indicated with the standard wavy red line. Words are stored in data/EnglishWords.txt, and you can add words to it if you want, one word per line (duplicates are ok).