Link hints and delete handling
FLASH links to files and directories now show their full paths in links when you pause your cursor over them.
And FLASH links now update correctly when files are deleted, where there are two or more paths to files with the same name. This includes repeated creating and deleting a file. To see the refreshed links, do a refresh of your browser tab, or make an edit if using the Editor.