New... button, and popup suppression
The Files page now has a "New..." button at top right, click it to make a new text or source file.
And glossary popups are now suppressed if they would overlap with a link, that is when a glossary term form all or part of a file name. In rare cases you might have to duplicate the term in your text to get the popup, for example instead of "...for more see "Trickystuff.txt" use "for more on Trickystuff see Trickystuff.txt".