QUICK INSTALL
For new installations, IntraMine now has a "quick installer" with instructions for running in IntraMine/__START_HERE_INTRAMINE_INSTALLER/1 READ ME FIRST how to install IntraMine.html. It's much easier than the old approach, at the price of not allowing a specific version of Elasticsearch or Universal ctags to be chosen.
Some minor "cosmetic" issues have also been cleared up here and there. For example, links to GitHub HTML documents in a regular .md Markdown file displayed with the Viewer or Editor will now go directly to a heading if one is provided.