Small fixes
Glossary popups on terms in the alt attribute in image links in Markdown files are now suppressed. A '+' can now be used in image link specifiers on the ToDo page. Fixed a docs typo. Added instructions to the error message if foldermonitor.ps1 is unresponsive. Maybe the least interesting update ever?