Faster Indexing, NAS link fix
Initial and all subsequent Elasticsearch indexing is now about twice as fast as before, on the order of 30 minutes for 250,000 files.
And links of the form \\SharedDrive\Folder... now open correctly in Search results. Using mapped drive letters instead of \\Computername\SharedFolder directly is still recommended, mainly to simplify accessing a share under a specific account.