Reserved start services
Put a count of 0 for a service in data/serverlist.txt and it won't start when IntraMine starts, but it will be available under the "Add one page server:" menu on the Status page. There won't be an entry for it in the top navigation bar or under "Page servers" on the Status page until it is started. It's "reserved" because a port number is reserved for it so there won't be any port conflict when you do start it. You'll find a complete list of available services in _copy_and_rename_to_data/serverlist.txt, including a few with a zero count.