Installer docs update
"__START_HERE_INTRAMINE_INSTALLER\1 READ ME FIRST how to install IntraMine.txt" and .html now mention that sometimes the Cmd window misbehaves and fails to interpret newlines, resulting in text running off the right of the window. And if this happens, the advice is to stop IntraMine and restart it.