Go to Anything, and smoother scrolling
Scrolling in views displayed by CodeMirror is now smoother, thanks to doing link marker redraws wrapped in startOperation/endOperation.

And IntraMine introduces Go to Anything: in the Viewer or Editor select a word, leave your cursor still, and a list of links will pop up. Could be the mention of a CSS selector in a JavaScript file, maybe a rare word in a text file, perhaps a JavaScript function that you'd like to visit - pretty much any single word anywhere. Now, that's a slight exaggeration, the usefulness of the resulting links depends on how rare the word is: you'll see up to eight links, and that's fine for a Go function or CSS selector but if it's a C++ method with 40 overrides then the links probably won't be the ones you want, alas.