Go2 and better startup
Go2: in the Viewer or Editor select a word or short phrase, leave your cursor fairly still, and a list of links will pop up. Could be the mention of a CSS selector in a JavaScript file, maybe a rare word in a text file, perhaps a JavaScript function that you'd like to visit - pretty much any single word or short phrase anywhere. Now, that's a slight exaggeration, the usefulness of the resulting links depends on how rare the word is: you'll see up to eight links, and that's fine for a Go function or a CSS selector but if it's a C++ method with 40 overrides then the links will probably be less useful. Only files in your indexed folders will be searched. These are the folders listed in data/search_directories.txt, that you initially set when you installed IntraMine. See Documentation/Go2.html for more.

And IntraMine feedback messages are no longer shown in its console window. Instead, they're shown on a new "Mon" page, listed in IntraMine's top navigation bar over on the right. By default a new fresh Mon page will come up in your default browser when you start IntraMine - to stop that from happening see Documentation/Mon.html.

There have been some fixes as well, in particular scrolling in views displayed by CodeMirror is now smoother.

And finally, Documentation has been revised so that each important feature in IntraMine has its own file.