Go2 and TOC upgrade
IntraMine no longer creates temporary files when calling universal ctags to list tags in a file, rather it uses Perl's backticks to pipe the tags straight into a variable. This is slightly faster, and removes the noise of temp files coming and going when Go2 is searching through multiple candidate files for instances of a word.
Affected: Go2, and table of contents creation for most languages (C++, JavaScript, Java etc).