Popup footnotes and citations
As with all of Gloss, put the markup for footnotes and citations in a .txt file and then view the converted HTML either by using IntraMine's Viewer or by converting to standalone HTML with the Glosser service (or the gloss2html.pl Perl program).

References look like[^27] for a footnote or [^27] for a citation (citations are preceded by a space). Footnote refs are shown superscripted, citation refs are shown as inline links.

And definitions for either look like
[^27]: this is the body of a footnote or citation. It continues until there's a blank line. Start flush left and don't forget the colon.

The complete footnote or citation ("note" for short) pops up when you pause your cursor over the reference link in the HTML version. As with glossary popups, this eliminates the need to jump away to view relevant information.

As opposed to glossary popups, notes are limited to a single file and a reference can be inserted anywhere. You can have multiple references to the same note.

Full details can be found at "Documentation/Gloss.html#Footnotes and citations".

And to correct some omissions that have been hanging around a bit:
 - Editor now has a proper table of contents when you're editing glossary files
 - Viewer now applies Gloss when displaying a glossary file
 - tooltips display more consistently when they appear one above the other on consecutive lines.