Better restart handling
When a service (or IntraMine as a whole) is restarted, who knows what should change on the front end? Could be some new JavaScript or CSS that wants loading into a page, or maybe the service port number for a page has changed due to deleting or reordering services in data/serverlist.txt. IntraMine now responds by completely reloading a page if its corresponding service restarts, adjusting the service port number if needed. This brings any IntraMine web page up to date. Note the Editor will not lose anything on a restart, though you will be advised if there are unsaved changes. Other pages left open such as Search, Files, or ToDo, will lose their state on a restart, but this is only a potential issue if you type an elaborate ToDo item and don't Save it before the restart. And that's an issue that is best avoided, for now, by not doing that.

On a related note, the "Search for" text box on the Search page now has a dropdown that will let you select from up to 10 recent search terms.