CSS table of contents fix
I forgot to correct an experiment with the CSS table of contents code. This has been fixed.
Also, if you hover over the file name at top left of an Editor view you will see a list of directories leading to the file: click on one to open the directory in IntraMine's Files page (or Windows File Explorer if the Files page isn't running).