Perl tidy, and a colon fix
All IntraMine Perl files have now been "ironed" with Perltidy. Maximum line length is 100, with a handful of strings and regular expressions running over (not sure if I want to break them up).

And colons are now properly escaped in all cases in glossary files (glossary_master.txt and glossary.txt), so you can put in an escaped colon with '\:' and it won't count as a term/definition separator.
(Glossaries are detailed in Documentation/Glossary popups.txt)