Search results fix, Go2 rizzed
Links in Search results were not working if there was a plus sign (+) in the path, which can happen with C++ enthusiasts. Fixed by applying encodeURIComponent more precisely.

And Go2 now tries harder: it no longer stops looking if it finds the defining instance of a term, but keeps looking for usages and mentions of the term. Definitions are shown first with a green background. This is subject to change, since a balance is needed between completeness of results and time taken to generate them.

Another fix, Go2 no longer generates a spurious 404 message in its command prompt window if nothing is found for a term or phrase.